88828b4 Fixed warnings about int-to-bool conversions.
6eaef7d Fixed warnings.

+++ Plugins/PointSprite/ParaViewPlugin/vtkPointSpriteRepresentation.cxx
@@ -240,7 +240,7 @@ void vtkPointSpriteRepresentation::SetRadiusArrayName(const char* val)
 
 //***************************************************************************
 // Forwarded to ScalarsToColorsPainter and LODScalarsToColorsPainter
-void vtkPointSpriteRepresentation::SetEnableOpacity(double val)
+void vtkPointSpriteRepresentation::SetEnableOpacity(int val)
 {
   this->ScalarsToColorsPainter->SetEnableOpacity(val);
   this->LODScalarsToColorsPainter->SetEnableOpacity(val);
